+++ Site/webapp/wdkCustomization/js/client/components/Visualizations/Table/TableFilters/FilterPanel.tsx
@@ -1,17 +1,12 @@
 // modified from https://github.com/ggascoigne/react-table-example
 
 import { Theme, createStyles, makeStyles } from "@material-ui/core";
-import Accordion from "@material-ui/core/Accordion";
-import AccordionDetails from "@material-ui/core/AccordionDetails";
-import AccordionSummary from "@material-ui/core/AccordionSummary";
-import AccordionActions from "@material-ui/core/AccordionActions";
-import Typography from "@material-ui/core/Typography";
-import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-import Chip from "@material-ui/core/Chip";
-import Button from "@material-ui/core/Button";
-import Divider from "@material-ui/core/Divider";
 
-import React, { FormEvent, ReactElement, useCallback } from "react";
+import Button from "@material-ui/core/Button";
+import Box from "@material-ui/core/Box";
+import Grid from "@material-ui/core/Grid";
+
+import React, { ReactElement, useCallback } from "react";
 import { TableInstance } from "react-table";
 
 const useStyles = makeStyles((theme: Theme) =>
@@ -55,7 +50,7 @@ const useStyles = makeStyles((theme: Theme) =>
         },
         grid: {
             display: "grid",
-            gridTemplateColumns: "repeat(2, 218px)",
+            gridTemplateColumns: "repeat(4, 218px)",
             "@media (max-width: 600px)": {
                 gridTemplateColumns: "repeat(1, 180px)",
             },
@@ -87,19 +82,14 @@ function FilterPanel({ instance }: FilterPageProps): ReactElement {
     }, [setAllFilters]);
 
     return (
-        <div className={classes.root}>
-            <Accordion>
-                <AccordionSummary expandIcon={<ExpandMoreIcon />} aria-controls="panel1c-content" id="panel1c-header">     
-                        <Typography className={classes.heading}>Advanced Filter</Typography>
-                </AccordionSummary>
-
-                <AccordionDetails className={classes.details}>
-                    <AccordionActions>
-                        <Button color="primary" onClick={resetFilters}>
-                            Clear Advanced Filters
-                        </Button>
-                    </AccordionActions>
-                    
+        <Box>
+            <Grid container direction="row">
+                <Grid item>
+                    <Button color="primary" onClick={resetFilters}>
+                        Clear Advanced Filters
+                    </Button>
+                </Grid>
+                <Grid item>
                     <form>
                         <div className={classes.grid}>
                             {/* pvalue */}
@@ -113,8 +103,11 @@ function FilterPanel({ instance }: FilterPageProps): ReactElement {
                                 ))*/}
                             {/* render pie charts first ; includes to handle special cases */}
                             {allColumns
-                                //@ts-ignore
-                                .filter((item) => item.canFilter && (item.filter && item.filter.toLowerCase().includes("pie")))
+                               
+                                .filter(
+                                     //@ts-ignore
+                                    (item) => item.canFilter && item.filter && item.filter.toLowerCase().includes("pie")
+                                )
                                 .map((column) => (
                                     <div key={column.id} className={classes.cell}>
                                         {column.render("Filter")}
@@ -135,10 +128,9 @@ function FilterPanel({ instance }: FilterPageProps): ReactElement {
                             &nbsp;
                         </Button>
                     </form>
-                </AccordionDetails>
-                <Divider />
-            </Accordion>
-        </div>
+                </Grid>
+            </Grid>
+        </Box>
     );
 }
