+++ Site/webapp/wdkCustomization/js/client/components/Visualizations/Table/TableTypes.tsx
@@ -9,17 +9,6 @@ import {
 
 export interface TableState<D extends object = {}> extends UsePaginationState<D> {}
 
-export interface CustomTableProps {
-    columns: Column<{}>[];
-    data: any;
-    canFilter: boolean;
-    filterTypes?: any; // json object of filter types
-    className?: string;
-    showAdvancedFilter?: boolean;
-    showHideColumns?: boolean;
-}
-
-
 export type TableMouseEventHandler = (instance: TableInstance) => MouseEventHandler;
 
 export type TableData = Record<string, string>;
